Add new script to the frontend project
This way, it's possible to test a production-like version of the
frontend, serving what's built of it.

Testing the build version of the app allowed me to ensure that only
one request is triggered to the API in the `useEffect` code, which
in development mode (`npm start`), runs twice purposefully.